+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added support for `TimedeltaIndex.mean` method.
 - Added support for some cases of aggregating `Timedelta` columns on `axis=0` with `agg` or `aggregate`.
 - Added support for `by`, `left_by`, and `right_by` for `pd.merge_asof`.
+- Added support for `resample` frequency `ME` with `closed = "left"`.
 
 #### Bug Fixes